chore: Update FaqResource toArray method to remove 'answer' field, change type column 'id' from INT to BIGINT

+++ database/migrations/2024_08_13_125027_create_faqs_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
 
         DB::statement(
             'CREATE TABLE faqs (
-                id INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+                id BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
                 question TEXT NOT NULL,
                 answer TEXT NOT NULL,
                 created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
